Added support for typesystem
Added typesystem model and editors
Added typesystem scoping: references to types now have
to be one of the pre defined options (Int, String, Bool,
Regex), or a reference to a user defined type in scope.
In scope means defined in the current file.

Added checks to make sure that names for user defined types are unique.
They cannot be the same as any other typname in the same file, or be one of the pre defined options.