Implemented basic expression transformation support.
Typing a binary operator creates the respective operator,
and assigns existing expressions to its left or right side depending on where it was typed.
Typing a brace and any other character creates a parenthesized expression node.
If the second character was a number, then that number is put inside the new parenthesized expression
Typing a number will now create a new constant expression node with that number in it

Creating any binary operator in any way will now automatically juggle the expression tree around until all precedence rules are satisfied.