Further work on expression system.
Added support for right associative operators.
Worked on system for restricting autocomplete to the right set of expressions in various cases.
Added styling for some behaviour expression constants.
Added support for typing behaviour expressions from left to right better.

Started work on scoping of channel names.
Currently, channels defined in the top of a modeldef need to be defined in a channeldef somewhere.
Channels used inside a modeldef have to be defined at the top of that modeldef.