Implemented comments.
Comments can now be added to a number of nodes.
Mostly definitions and single line behaviour expressions
are supported.
Comments can be added or removed with alt + enter.
Alternatively, they can also be added or removed with
cntrl + shift + /. This second keybind looks for the closest
commentable ancestor.
Both single line and multi line comments are supported.
Comments are also added to the generated output.