Removes typevariable as a type
Typelabels now use node<> instead of node<type> as their typearguments.
Runtimetypevariables are now being used directly, without the typevariable wrapper
making them an implementation of the type interface.
Unboundgeneric type is added as a new type. It represents having knowledge of the type of a node,
but not knowing the bindings for the generics of that type.
an unboundgeneric type with a name is a supertype of a typelabel with the same name, which will also
add the information of the generic bindings.